requirement.txt
django-lint==2.0.4
Django==4.1.2
djangorestframework==3.14.0
python-decouple==3.6
drf-yasg==1.21.5
django-extensions==3.2.1
psycopg2==2.9.5
pytest==7.1.2